.agents/skills/ocr/scripts/requirements.txt
# Core dependencies (DeepSeek-OCR)
requests>=2.28.0

# PDF support
pdf2image>=1.16.0
Pillow>=9.0.0

# Optional: Fast mode (PaddleOCR)
# paddleocr>=3.4.0
# paddlepaddle>=3.0.0
